requirements.txt
# ─── Core Framework ───
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0

# ─── LLM (Local) ───
llama-cpp-python>=0.3.0

# ─── Embeddings ───
sentence-transformers>=3.3.0
FlagEmbedding>=1.2.0
torch>=2.1.0
transformers>=4.40.0

# ─── Vector Store ───
faiss-cpu>=1.9.0

# ─── Sparse Retrieval ───
rank-bm25>=0.2.2

# ─── Document Processing ───
tiktoken>=0.8.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
markdown>=3.7

# ─── Auth & Security ───
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# ─── Observability ───
structlog>=24.0.0
prometheus-client>=0.21.0

# ─── Scientific / Metrics ───
numpy>=1.24.0
scikit-learn>=1.5.0
pandas>=2.0.0

# ─── Testing ───
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
